Address review comments
* Change Decimal to float
* Move dataclasses to separate charging_profile module
* Add ChargingProfile type to RemoteStartTransaction payload
